Include immintrin.h headers when using clang-cl.
clang-cl defines _MSC_VER which when including <immintrin.h>
causes all the headers for instruction sets not already defined in
the compile line to be excluded. This means that when compiling for
dynamic dispatch targets the required headers are not actually included
when using clang-cl. This patch worksaround the issue by directly
including those targets in that case.