Replace %zu printf usage with uint64_t ones.
%zu is not supported everywhere, in particular it is not supported in
MSYS2. Most of the current uses were in unittests, but some public
headers also had some meaning that projects using hwy would hit those
occurrences.

This patch replaces all of these usages in most cases replacing them by
uint64_t, which might be a larger type in 32-bit platforms but that's
not an issue. %zu is only used in combination with printf() to stdout or
stderr in cases where performance is not an issue.