Defer the call to get timer resolution until needed
We have a use case where the highway library is linked into a shared object.
Currently whenever it is dlopen'ed, the static initializer runs including
running the `platform::TimerResolution`. This has been observed in profiling and
raised some questions, namely it shouldn't be necessary to initialize the
benchmarking code if we are using the highway library but not to benchmark
anything.

Furthermore, doing so is against many companies' C++ guidelines, making this a
blocker for inclusion into companies' code bases.

This commit changes this variable from a namespace-scope static variable into a
static local variable. It may have a little bit more overhead but it should be
negligible.