Dispatch LSX/LASX at runtime
Currently the CMake and meson build systems of highway silently appends
-mlasx to compiler flags.  But the Software Development and Build
Convention for LoongArch Architectures manual only states the LoongArch
desktop and server processors should support LSX, not LASX.  And there
is now at least one LoongArch desktop processor model which lacks LASX,
namely Loongson 3B6000M.

Silently appending -mlasx to flags is also very unfriendly to the
downstream builders like distro maintainers: if they'd like to make a
distro specifically optimized with LASX (and give up the compatibility
with less-capable processors) they'd make -mlasx the default of their
toolchain or packaging system anyway so we don't need to appending it;
but if they want to be compatible with less-capable processors like
3B6000M this behavior will just surprise them by triggering some SIGILL
at runtime and requiring special build system option to be turned off.

To take the advantage of LASX, it's better to make it dispatched at
runtime.  To achieve that we need to utilize a pragma provided by the
compiler, similar to how we implement it for x86 and aarch64.
Unfortunately, the target pragma support of LoongArch GCC is only
recently introduced in GCC 15, and it's still not enough to support
highway due to https://gcc.gnu.org/PR121875.  So the runtime dispatch
is only implemented for Clang >= 18 (the minimum version with LSX and
LASX support) as at now.  The GCC support can be added once the GCC
issue is fixed.